+++ tests/unit/lxd/test_lxc.py
@@ -1006,11 +1006,25 @@ def test_launch_failed_retry_check(fake_process, mocker):
     ]
 
 
+@pytest.mark.parametrize("has_stderr", [True, False])
+@pytest.mark.parametrize("state", ["create", "start"])
 @pytest.mark.usefixtures("mock_getpid")
-def test_launch_failed_retry_failed(fake_process, mocker):
-    """Test that we retry launching an instance if it fails, but failed more than 3 times."""
+def test_launch_failed_retry_failed_in_progress(
+    fake_process, mocker, has_stderr, state
+):
+    """Retry launching 3 times if another process is launching the instance."""
+    if has_stderr:
+        stderr = (
+            "Error: Failed instance creation: Failed creating instance record: "
+            f'Instance is busy running a "{state}" operation'
+        ).encode()
+    else:
+        stderr = None
+
     mock_launch = mocker.patch("craft_providers.lxd.lxc.LXC._run_lxc")
-    mock_launch.side_effect = subprocess.CalledProcessError(1, ["lxc", "fail", " test"])
+    mock_launch.side_effect = subprocess.CalledProcessError(
+        1, ["lxc", "fail", " test"], stderr=stderr
+    )
     mock_check = mocker.patch("craft_providers.lxd.lxc.LXC.check_instance_status")
     mock_check.side_effect = LXDError("test")
     mocker.patch("time.sleep")
@@ -1068,6 +1082,50 @@ def test_launch_failed_retry_failed(fake_process, mocker):
             capture_output=True,
             project="test-project",
         ),
+        call(
+            [
+                "launch",
+                "test-image-remote:test-image",
+                "test-remote:test-instance",
+                "--config",
+                "user.craft_providers.status=STARTING",
+                "--config",
+                "user.craft_providers.timer=2023-01-01T00:00:00+00:00",
+                "--config",
+                "user.craft_providers.pid=123",
+            ],
+            capture_output=True,
+            stdin=StdinType.INTERACTIVE,
+            project="test-project",
+        ),
+    ]
+
+
+@pytest.mark.usefixtures("mock_getpid")
+def test_launch_failed_retry_failed(fake_process, mocker):
+    """Fail quickly the error isn't related to another process launching the instance."""
+    mock_launch = mocker.patch("craft_providers.lxd.lxc.LXC._run_lxc")
+    mock_launch.side_effect = subprocess.CalledProcessError(
+        1,
+        ["lxc", "fail", " test"],
+        # use an error that doesn't indicate another process is launching the instance ('create' or 'start' is not in the error)
+        stderr=b"test error",
+    )
+    mock_check = mocker.patch("craft_providers.lxd.lxc.LXC.check_instance_status")
+    mock_check.side_effect = LXDError("test")
+    mocker.patch("time.sleep")
+
+    with pytest.raises(LXDError):
+        with freeze_time("2023-01-01"):
+            LXC().launch(
+                instance_name="test-instance",
+                image="test-image",
+                image_remote="test-image-remote",
+                project="test-project",
+                remote="test-remote",
+            )
+
+    assert mock_launch.mock_calls == [
         call(
             [
                 "launch",