+++ docs/reference/changelog.rst
@@ -4,6 +4,18 @@ Changelog
 See the `Releases page`_ on GitHub for a complete list of commits that are
 included in each version.
 
+2.3.0 (upcoming)
+----------------
+
+New features:
+
+- Add support for Ubuntu 25.04 (Plucky).
+
+Bug fixes:
+
+- Fix a race condition where multiple processes would fail to launch LXD
+  instances at the same time.
+
 2.2.0 (2025-Feb-11)
 -------------------